moved some tests to correct file
